Edits fixing the main CAVA program
* On root directory run `2to3 -w ./` which will fix most of the
  incompatiblities. 2to3 ships with python

* Open gzip files in `rt` mode [`gzip.open(<file.gz>), 'rt'`] so that
  each line is returned as string instead of as binary

* Integer division is explicitly cast to integer where an integer is
  expected

* Major version bump to 2.0.0

+++ cava_/csn.py
@@ -4,7 +4,7 @@
 # CSN annotation
 #######################################################################################################################
 
-import core
+from . import core
 
 
 # Class representing a CSN annotation
@@ -175,7 +175,7 @@ def makeProteinString(variant, transcript, reference, prot, mutprot, coord1):
 
     # Checking if there was no change in protein sequence
     if prot == mutprot:
-        idx = coord1 / 3
+        idx = int(coord1 / 3)
         if coord1 % 3 > 0: idx += 1
         return '_p.=', (str(idx), prot[idx-1], prot[idx-1])
 
@@ -281,7 +281,7 @@ def transformToCSNCoordinate(pos, transcript):
                 if transcript.strand == 1:
                     # Checking if genomic position is within intron
                     if prevExonEnd < pos < exon.start + 1:
-                        if pos <= (exon.start + 1 - prevExonEnd) / 2 + prevExonEnd:
+                        if pos <= int((exon.start + 1 - prevExonEnd) / 2) + prevExonEnd:
                             x, y = transformToCSNCoordinate(prevExonEnd, transcript)
                             return x, pos - prevExonEnd
                         else:
@@ -290,7 +290,7 @@ def transformToCSNCoordinate(pos, transcript):
                 else:
                     # Checking if genomic position is within intron
                     if exon.end < pos < prevExonEnd:
-                        if pos >= (prevExonEnd - exon.end + 1) / 2 + exon.end:
+                        if pos >= int((prevExonEnd - exon.end + 1) / 2) + exon.end:
                             x, y = transformToCSNCoordinate(prevExonEnd, transcript)
                             return x, prevExonEnd - pos
                         else:
@@ -320,7 +320,7 @@ def transformToCSNCoordinate(pos, transcript):
                 if transcript.strand == 1:
                     # Checking if genomic position is within intron
                     if prevExonEnd < pos < exon.start + 1:
-                        if pos <= (exon.start + 1 - prevExonEnd) / 2 + prevExonEnd:
+                        if pos <= int((exon.start + 1 - prevExonEnd) / 2) + prevExonEnd:
                             x, y = transformToCSNCoordinate(prevExonEnd, transcript)
                             return x, pos - prevExonEnd
                         else:
@@ -329,7 +329,7 @@ def transformToCSNCoordinate(pos, transcript):
                 else:
                     # Checking if genomic position is within intron
                     if exon.end < pos < prevExonEnd:
-                        if pos >= (prevExonEnd - exon.end + 1) / 2 + exon.end:
+                        if pos >= int((prevExonEnd - exon.end + 1) / 2) + exon.end:
                             x, y = transformToCSNCoordinate(prevExonEnd, transcript)
                             return x, prevExonEnd - pos
                         else:
